Fixing Issues 2+3 (#4)
* First patch

* Revised Patch 1

* Patch 1.2

Deleted the settings.json plus various other changes

+++ modmail/bot/bot.py
@@ -54,14 +54,22 @@ class Bot(commands.Bot):
         self.scheduler.start()
         print(f" Scheduler started ({len(self.scheduler.get_jobs())} jobs scheduled).")
 
-        await self.change_presence(activity=discord.Activity(name=f"DM reports • Version {modmail.__version__}", type=discord.ActivityType.listening))
+        await self.change_presence(
+            activity=discord.Activity(
+                name=f"DM reports • Version {modmail.__version__}", type=discord.ActivityType.listening
+            )
+        )
         print(f" Presence set.")
 
         self.guild = self.get_guild(Config.GUILD_ID)
+        self.staff_role = self.guild.get_role(Config.STAFF_ROLE_ID)
         print(f" Bot ready.")
 
     async def on_message(self, message: discord.Message) -> None:
         if message.guild and message.guild.me in message.mentions:
+            if self.staff_role in message.author.roles:
+                return
+
             await message.delete()
             try:
                 await message.author.send(
